% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{gcap.enrich}
\alias{gcap.enrich}
\title{Enrichment Analysis}
\usage{
gcap.enrich(
  geneList,
  analysis_func = c("enricher", "fgsea"),
  gene_encode = c("ensembl", "symbol"),
  species = "Homo sapiens",
  category = "H",
  subcategory = "",
  genome_build = c("hg38", "hg19"),
  ...
)
}
\arguments{
\item{geneList}{a vector (list) of genes.
If the vector is named, the names must be genes and values must numeric.}

\item{analysis_func}{analysis function to run.}

\item{gene_encode}{gene ID type from input.}

\item{species}{species, not to change.}

\item{category}{gene set main category name.}

\item{subcategory}{gene set sub-category name.}

\item{genome_build}{genome version.}

\item{...}{other parameters passing to \code{clusterProfiler::enricher()}
or \code{fgsea::fgsea}.}
}
\value{
depends on \code{data} and \code{analysis_func}.
}
\description{
Enrichment Analysis
}
\examples{
\donttest{
if (require("clusterProfiler")) {
  gcap.enrich(c("TP53", "MYC", "ABC", "EGFR", "GSX2"), gene_encode = "symbol")
}

if (require("fgsea")) {
  gcap.enrich(c("TP53", "MYC", "ABC", "EGFR", "GSX2"),
    gene_encode = "symbol", analysis_func = "fgsea"
  )
}
}
}
