% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{gcap.plotCircos}
\alias{gcap.plotCircos}
\title{Plot Circos}
\usage{
gcap.plotCircos(
  fCNA,
  highlights = NULL,
  clust_distance = 1e+07,
  col = c("#FF000080", "#0000FF80"),
  genome_build = c("hg38", "hg19"),
  chrs = paste0("chr", 1:22),
  ideogram_height = 1,
  prob_cutoff = 0.5,
  gap_cn = 4L,
  ...
)
}
\arguments{
\item{fCNA}{a \href{gcap::fCNA}{fCNA} object.}

\item{highlights}{gene/cytoband list to highlight.}

\item{clust_distance}{a distance as cutoff for different clusters.
Default is 1e7, i.e. 10Mb. Note 100 Mb is set to genes on different
chromosomes, so please don't set value larger than that.}

\item{col}{length-2 colors for circular and noncircular.}

\item{genome_build}{genome version.}

\item{chrs}{chromosome names.}

\item{ideogram_height}{ideogram height at default.}

\item{prob_cutoff}{thresholds for classify 'circular' and 'noncircular'.}

\item{gap_cn}{thresholds for classify 'circular' and 'noncircular'.}

\item{...}{other parameters passing to \code{ggrepel::geom_label_repel()}.}
}
\value{
Nothing.
}
\description{
Plot Circos
}
\examples{
\donttest{
library(gcap)
if (require("circlize") && require("scales")) {
  data("ascn")
  data <- ascn

  # Create fake data
  set.seed(1234)
  data$sample <- sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  gcap.plotCircos(rv)

  # Select genes to highlight in plot
  r1 <- rv$getGeneSummary(return_record = TRUE)
  gcap.plotCircos(rv, r1[amplicon_type == "circular"]$gene_id[1:10])

  gcap.plotCircos(rv, r1[, .(label = .N), by = .(band)][1:10])
  gcap.plotCircos(
    rv,
    r1[, .(label = .N, cluster = gsub("(.*):.*", "\\\\1", band)),
      by = .(band)
    ][1:10]
  )
}
}
}
