% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{gcap.plotDistribution}
\alias{gcap.plotDistribution}
\title{Draw Gene Amplicon Distribution By User Specified Class}
\usage{
gcap.plotDistribution(
  fCNA,
  x = NULL,
  x_size = 8,
  set_order = TRUE,
  set_label = TRUE,
  fill = TRUE,
  palette = c("#CCCCCC", "#0066CC", "#FFCCCC", "#CC0033"),
  ...
)
}
\arguments{
\item{fCNA}{a \code{fCNA} object or a \code{data.frame} with at least
3 columns "sample", "class" and "by".}

\item{x}{a column name in \code{fCNA$sample_summary}.}

\item{x_size}{font size of x axis text.}

\item{set_order}{set order for \code{by}.}

\item{set_label}{set labels to add sample count for \code{by}.}

\item{fill}{if \code{TRUE}, show the percentage instead of count.}

\item{palette}{color palette.}

\item{...}{other parameters passing to \code{ggplot2::geom_bar}.}
}
\value{
a ggplot object.
}
\description{
Draw Gene Amplicon Distribution By User Specified Class
}
\examples{
set.seed(1234)
data = data.frame(
  sample = sample(LETTERS[1:10], 100, replace = TRUE),
  class = sample(c("nofocal", "noncircular", "circular"), 100, replace = TRUE),
  by = sample(1:4, 100, replace = TRUE)
)
p = gcap.plotDistribution(data)
p
}
