% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{gcap.plotProfile}
\alias{gcap.plotProfile}
\title{Draw Oncoprint for fCNA Profile Visualization}
\usage{
gcap.plotProfile(
  data,
  genes = NULL,
  samples = NULL,
  top_n = NULL,
  show_column_names = TRUE,
  remove_empty_columns = FALSE,
  remove_empty_rows = TRUE,
  heatmap_legend_param = list(title = "fCNA"),
  col = c(noncircular = "#0066CC", circular = "#CC0033"),
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} to draw oncoprint, rows represent
genes/cytobands and columns represent samples.}

\item{genes}{a list of genes. Must be subsets of \code{fCNA$data$gene_id}.}

\item{samples}{a vector to subset samples.}

\item{top_n}{top N genes to show.}

\item{show_column_names}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{remove_empty_columns}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{remove_empty_rows}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{heatmap_legend_param}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{col}{see \code{?ComplexHeatmap::oncoPrint}.}

\item{...}{Other parameters passing to \code{ComplexHeatmap::oncoPrint}.}
}
\value{
a \code{Heatmap} object, you can go further modify it.
}
\description{
See \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/oncoprint.html}{oncoprint}
and \href{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#size-of-the-heatmap}{heatmap}
to modify the result plot in depth.
}
\examples{
\donttest{
library(gcap)
if (require("ComplexHeatmap") && require("IDConverter")) {
  data("ascn")
  data = ascn

  # Create fake data
  set.seed(1234)
  data$sample = sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv = gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  data2 = rv$getGeneSummary(return_mat = TRUE)
  gcap.plotProfile(data2)

  rv$convertGeneID()
  data2 = rv$getGeneSummary(return_mat = TRUE)
  ht = gcap.plotProfile(data2,
    samples = c("B", "A", "C", "D", "F"),
    genes = rownames(data2)[1:10],
    top_annotation = ComplexHeatmap::HeatmapAnnotation(
      cbar = ComplexHeatmap::anno_oncoprint_barplot(),
      foo1 = c("g1", "g1", "g2", "g2", "g3"),
      bar1 = ComplexHeatmap::anno_points(1:5)
    )
  )
  ht
  ComplexHeatmap::draw(ht, merge_legends = TRUE)

  data2 = rv$getCytobandSummary(return_mat = TRUE)
  gcap.plotProfile(data2)
}
}
}
\seealso{
\link[gcap:fCNA]{gcap::fCNA} for building object.
}
