% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcaputils-package.R
\docType{package}
\name{gcaputils-package}
\alias{gcaputils}
\alias{gcaputils-package}
\title{gcaputils: Utils for Package 'gcap'}
\description{
Auxiliary tools for analysis and visualization of focal copy number amplification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShixiangWang/gcaputils}
  \item Report bugs at \url{https://github.com/ShixiangWang/gcaputils/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
