% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distHeatmap.R
\name{gcap.plotGenomeHeatmap}
\alias{gcap.plotGenomeHeatmap}
\title{Plot Genome-level Heatmap}
\usage{
gcap.plotGenomeHeatmap(
  fCNA,
  group = NULL,
  type = c("class", "prob", "circ_freq", "noncirc_freq", "circ_cn", "noncirc_cn"),
  highlights = NULL,
  total_n = NULL,
  add_all = TRUE,
  draw = TRUE,
  bycol = TRUE,
  sort = FALSE,
  w = 1e+06,
  title = "auto",
  title_rot = 90,
  fontsize = 8,
  top_col = "red",
  max_val = NULL,
  genome_build = c("hg38", "hg19"),
  chrs = paste0("chr", 1:22)
)
}
\arguments{
\item{fCNA}{a \href{gcap::fCNA}{fCNA} object.}

\item{group}{a group column for comparison, default is "sample".}

\item{type}{data type to plot.}

\item{highlights}{gene list or region (in \code{data.frame}) to highlight.}

\item{total_n}{if not \code{NULL}, this is used for calculating percentage.}

\item{add_all}{if \code{TRUE}, add freq barplot for noncircular and circular amplicons.
This works better with \code{type = "prob"}.}

\item{draw}{if \code{TRUE}, draw plot.}

\item{bycol}{if \code{FALSE}, draw in row manner.}

\item{w}{window size.}

\item{title}{title for heatmap legend.}

\item{title_rot}{rotation of the title.}

\item{fontsize}{fontsize for highlight.}

\item{top_col}{color setting.}

\item{max_val}{maximum value in heatmap.}

\item{genome_build}{genome version.}

\item{chrs}{chromosome names.}
}
\value{
Nothing.
}
\description{
If \code{highlight} is set, make sure use \code{pdf()} or other device for
rendering plot.
}
\examples{
\donttest{
library(gcap)
if (require("circlize") && require("scales")) {
  data("ascn")
  data <- ascn

  # Create fake data
  set.seed(1234)
  data$sample <- sample(LETTERS[1:10], nrow(data), replace = TRUE)
  rv <- gcap.ASCNworkflow(data, outdir = tempdir(), model = "XGB11")

  gcap.plotGenomeHeatmap(rv,
    highlights = c("TP53", "EGFR", "ERBB2"),
    add_all = TRUE
  )
}
}
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/genome-level-heatmap.html
}
